# go api: reduce allocations in deck decompression